now works correctly in the shadow window
the secret of the misunderstanding here is
that chipmunk does NOT assume constant
density. it simply uses "center of gravity"
as a term for "the local origin".

The center of gravity of a body is
by definition at (0,0).

(However, the time spent calculating the
centroids was not wasted, since this
indicates how we should position the
bodies so they will look natural.)