now moments are important!
should calculate this properly, but adjusted
it up for feel.

(the rotational force must be matched with
the expected ship moments to provide the
required behaviour, effectively nullifying
the direct effect of the moment.)

However, in the event of a collision
angular momentum (not speed!) is important.
Try adjusting the moment and sending some
boxes flying by rotation.