promising performance with many offscreen objs
also note that I'm still testing on Mesa..

with all these 5000 objects on the screen at
once, the FPS is something like 160 (which, it
should be noted, is still plenty fast).

5000 objects on the screen at once is good as
high-stress test. we'll have screen space for
around 1000 objects on the screen at once
(remember they will not overlap), and we can
probably expect around 500 in busy moments.

however, we are currently not dealing with
the major aspect of updates to positions
and angles based on the physics engine.