refactor + experiment with new projectile type
bug uncovered: as far as sprites are concerned
projectiles are spawned with their _middle_
over the gun, instead of outside it.
does this mesh with the physics?

(might actually not be a bug, could be they
should just have z below the ships, so they
appear to emerge gradually from the gun)